Combine wayfinding to core plugin
