+++ src/android/CurrentStatus.java
