--- src/android/IndoorAtlasWayfinding/IACordovaWayfinding.gradle
@@ -1,11 +1,0 @@
-repositories {
-    mavenCentral()
-    maven {
-        url "https://indooratlas-ltd.bintray.com/mvn-public"
-    }
-}
-
-dependencies {
-    compile 'com.android.support:appcompat-v7:27.0.2'
-    compile 'com.indooratlas.android:indooratlas-android-wayfinding-sdk:0.1.1-dev-5@aar'
-}
