+++ src/ios/IndoorAtlas/IndoorAtlasWayfinding.framework/Headers/wayfinding.h
