+++ src/ios/IndoorAtlas/IndoorAtlasWayfinding.framework/Info.plist
