+++ src/ios/IndoorAtlas/IndoorAtlasWayfinding.framework/Modules/module.modulemap
