+++ src/ios/IndoorAtlasLocationService.h
