+++ src/ios/IndoorAtlasLocationService.m
